More quickly detect a daemonset is up and running
Overall I see the daemonsets on my nodes be ready in approx 5 to 7 seconds.
Reducing the initial delay allows them to be marked as ready within 10 seconds. In cases it takes longer the next probe will be in 30 seconds so max initial waiting time is 40 seconds to be marked as ready compared to 75 seconds at this point of time. Furhtermore the 30 seconds probes also allow more quick detection of failures

Signed-off-by: Marco Franssen <marco.franssen@gmail.com>